First push and Queue with Two Stacks
The associated files are the result of the first push of the HackerRankSolutions project, and introduction of the 1st-attempt solution code for making a queue with two stacks.

Problems with Queue with Two Stacks current form: conceptual soundness, but test cases 6-15 of 16 test cases experience timeout; ergo, greater efficiency is required for the final solution.