Queue with Two Stacks efficiency improvements
Fixed minor typographical error in problem transcript.

Alterations to solution code:
Upon reading about other solution approaches, a new strategy was adopted based on a more efficient working algorithm. The content in one stack is no longer left as-is until a "dequeue" or "peek" operation is needed and then boxing the contents back and forth between the two stacks. Instead the contents of one stack are moved to the other whenever the second stack is empty and a "dequeue" or "peek" is needed, and the second stack's top element is removed or examined; until the contents of the second stack are exhausted, it is unnecessary to move any more elements over from the first stack.