Down to Zero II
Added working solution for Down to Zero II problem; following repeated attempts to implement a suggested solution, several modifications were required. These include placeholder calculations for a 0-step operation and an alternative method for finding the appropriate number of steps for a given input (the latter is suspected as the problem root in the prior, unpushed implementation).