Game of Two Stacks fixes
Alterations to solution code:

As a housekeeping measure, the Scanner object used for reading inputs is now manually closed.

While the principles of the previous code may have been sound in most cases, the code also did not take into account the times when stack elements might have a value of 0. These instances led to circumstances where, for example, a single push/pop was made when multiple operations were appropriate, and vice-versa. This extended to instances where a value of 0 caused loops to execute needlessly, and thus circumstantially execute longer than permitted.